components/dataset-splitter/requirements.txt
google-cloud-storage
PyYAML